Rag_minimal_day8/gateway/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pyyaml==6.0.1
requests==2.31.0